IBC loading fix Commit to rebuild core on x86

+++ x/oracle/keeper/result.go
@@ -51,7 +51,7 @@ func (k Keeper) MustGetResult(ctx sdk.Context, id oracletypes.RequestID) oraclet
 }
 
 // ResolveSuccess resolves the given request as success with the given result.
-func (k Keeper) ResolveSuccess(ctx sdk.Context, id oracletypes.RequestID, result []byte, gasUsed uint32) {
+func (k Keeper) ResolveSuccess(ctx sdk.Context, id oracletypes.RequestID, result []byte, gasUsed uint64) {
 	k.SaveResult(ctx, id, oracletypes.RESOLVE_STATUS_SUCCESS, result)
 	ctx.EventManager().EmitEvent(sdk.NewEvent(
 		oracletypes.EventTypeResolve,
